Update to work with Python 3
The Hamster Wheel of Backwards Incompatibility turns ever onward.  This probably
breaks support for Python 2, but at this point I don't care.  I just can't
summon up the willpower to do anything more than the most basic unfucking of
this thing that's been finished for 11 years but still breaks constantly because
of Mercurial's API churn and Python 3 setting the world on fire.

                                                                                    The
_  _ ____ _  _ ____ ___ ____ ____    _ _ _ _  _ ____ ____ _       ____ ____    ___  ____ ____ _  _ _ _ _ ____ ____ ___  ____    _ _  _ ____ ____ _  _ ___  ____ ___ _ ___  _ _    _ ___ _   _
|__| |__| |\/| [__   |  |___ |__/    | | | |__| |___ |___ |       |  | |___    |__] |__| |    |_/  | | | |__| |__/ |  \ [__     | |\ | |    |  | |\/| |__] |__|  |  | |__] | |    |  |   \_/
|  | |  | |  | ___]  |  |___ |  \    |_|_| |  | |___ |___ |___    |__| |       |__] |  | |___ | \_ |_|_| |  | |  \ |__/ ___]    | | \| |___ |__| |  | |    |  |  |  | |__] | |___ |  |    |
                                                                                turns forever